ecommerce_backend/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
pillow==10.4.0
platformdirs==4.2.2
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.5.1
virtualenv==20.26.3
